% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_WL.R
\name{initialize_WL}
\alias{initialize_WL}
\title{Initialize the shared sub-matrix E and modality specific sub-matrices list Hi}
\usage{
initialize_WL(WL, k)
}
\arguments{
\item{WL}{a list of multiple modality data matrices}

\item{k}{the cluster number}
}
\value{
res, a list of length N+3, where N is the number of data modality. the first N elements are the modality specific sub-matrices list Hi, the (N+1) element is the shared sub-matrix E, the last two elements are the loss defined on the shared sub-matrix E and modality specific sub-matrices list Hi.
}
\description{
Initialize the shared sub-matrix E and modality specific sub-matrices list Hi
}
\examples{
library(InterSIM)
sim.data <- InterSIM(n.sample=500, cluster.sample.prop = c(0.20,0.30,0.27,0.23),
delta.methyl=5, delta.expr=5, delta.protein=5,p.DMP=0.2, p.DEG=NULL,
p.DEP=NULL,sigma.methyl=NULL, sigma.expr=NULL, sigma.protein=NULL,cor.methyl.expr=NULL,
cor.expr.protein=NULL,do.plot=FALSE, sample.cluster=TRUE, feature.cluster=TRUE)
sim.methyl <- sim.data$dat.methyl
sim.expr <- sim.data$dat.expr
sim.protein <- sim.data$dat.protein
temp_data <- list(sim.methyl, sim.expr, sim.protein)
init_list <- initialize_WL(temp_data,k=4)
}
