% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersim_data_gen.R
\name{intersim_data_gen}
\alias{intersim_data_gen}
\title{Generate the simulated dataset with three modalities with the package InterSIM}
\usage{
intersim_data_gen(prop=c(0.20,0.30,0.27,0.23), n_sample=500)
}
\arguments{
\item{prop}{proportion of samples for each cluster}

\item{n_sample}{the number of samples}
}
\value{
res, a list of length 2, where the first element is a list of simulated data, while the second element is a vector indicating the true label of each sample.
}
\description{
Generate the simulated dataset with three modalities with the package InterSIM
}
\examples{
library(InterSIM)
intersim_data <- intersim_data_gen(prop=c(0.20,0.30,0.27,0.23), n_sample=500)
}
